ai_engine/requirements.txt
fastapi
uvicorn
openai
apscheduler
python-dotenv
supabase
requests
